@@ -1,5 +1,6 @@
-﻿using Animation2Tilemap.Services;
-using Animation2Tilemap.Services.Contracts;
+﻿using Animation2Tilemap.Core;
+using Animation2Tilemap.Core.Services;
+using Animation2Tilemap.Core.Services.Contracts;
 using Animation2Tilemap.Test.TestHelpers;
 using Moq;
 using Serilog;
@@ -8,21 +9,14 @@ using Xunit.Abstractions;
 
 namespace Animation2Tilemap.Test.Services;
 
-public class ImageLoaderServiceTests
+public class ImageLoaderServiceTests(ITestOutputHelper testOutputHelper)
 {
-    private readonly Mock<IConfirmationDialogService> _confirmationDialogServiceMock;
-    private readonly Mock<INamePatternService> _namePatternServiceMock;
-    private readonly Logger _logger;
-
-    public ImageLoaderServiceTests(ITestOutputHelper testOutputHelper)
-    {
-        _logger = new LoggerConfiguration()
-            .MinimumLevel.Verbose()
-            .WriteTo.Sink(new TestOutputHelperSink(testOutputHelper))
-            .CreateLogger();
-        _confirmationDialogServiceMock = new Mock<IConfirmationDialogService>();
-        _namePatternServiceMock = new Mock<INamePatternService>();
-    }
+    private readonly Mock<IConfirmationDialogService> _confirmationDialogServiceMock = new();
+    private readonly Logger _logger = new LoggerConfiguration()
+        .MinimumLevel.Verbose()
+        .WriteTo.Sink(new TestOutputHelperSink(testOutputHelper))
+        .CreateLogger();
+    private readonly Mock<INamePatternService> _namePatternServiceMock = new();
 
     [Fact]
     public void TryLoadImages_ShouldReturnFalse_WhenInputPathIsInvalid()
@@ -140,7 +134,7 @@ public class ImageLoaderServiceTests
         // Assert
         Assert.True(result);
         Assert.Equal(8, images["anim"].Count);
-        Assert.Equal(1, images.Count);
+        Assert.Single(images);
 
         foreach (var image in images["anim"])
         {
